requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
